fix: barcodereader now correctly reads data
